requirements.txt
asgiref==3.10.0
Django==5.2.7
django-extensions==4.1
djangorestframework==3.16.1
django-cors-headers==4.6.0
mysqlclient==2.2.7
sqlparse==0.5.3
tzdata==2025.2
